delete all images corresponding to user and item

+++ userhandles.go
@@ -231,19 +231,30 @@ func (cb *Corkboard) DeleteUser(w http.ResponseWriter, r *http.Request, ps httpr
 		return
 	}
 
+	cb.deleteImageID(theuser.PicID) //nolint: errcheck
+
 	_, err := cb.Bucket.Remove(key, 0)
 	if err != nil {
 		log.Println(err)
 		return
 	}
 
-	items, _ := cb.findUserItems(id)
+	//Check if user has any existing items, if so delete them
+	items, err := cb.findUserItems(id)
+	if len(items) != 0 && err == nil {
 
-	for i := 0; i < len(items); i++ {
-		var docID = "item:" + items[i].ID
-		_, err := cb.Bucket.Remove(docID, 0)
-		if err != nil {
-			log.Println(err)
+		for i := 0; i < len(items); i++ {
+
+			//delete images for each picture
+			for j := 0; j < len(items[i].Pic); j++ {
+				cb.deleteImageID(items[j].Pic) //nolint: errcheck
+			}
+
+			var docID = "item:" + items[i].ID
+			_, err := cb.Bucket.Remove(docID, 0)
+			if err != nil {
+				log.Println(err)
+			}
 		}
 	}
